requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
xgboost>=2.0.0
catboost>=1.2.0
scipy>=1.11.0
openpyxl>=3.1.0
joblib>=1.3.0
pyarrow>=14.0.0
umap-learn>=0.5.5
hdbscan>=0.8.33
requests>=2.31.0
